+++ examples/e2e.py
@@ -109,7 +109,8 @@ async def send_file(connection):
     message = FileMessage(
         connection=connection,
         to_id='ECHOECHO',
-        file_path='res/some_file.zip'
+        file_path='res/some_file.zip',
+        caption="Here's that file I mentioned",
     )
     return await message.send()