+++ threema/gateway/e2e.py
@@ -1138,6 +1138,7 @@ class FileMessage(Message):
           be used instead of passing the key directly.
         - `file_path`: The path to a file.
         - `thumbnail_path`: The path to a thumbnail of the file.
+        - `caption`: Caption of the file.
 
     Arguments for an existing message:
         - `payload`: The remaining byte sequence of a decrypted
@@ -1154,7 +1155,7 @@ class FileMessage(Message):
     def __init__(
             self, connection,
             file_content=None, mime_type=None, file_name='file', file_path=None,
-            thumbnail_content=None, thumbnail_path=None,
+            thumbnail_content=None, thumbnail_path=None, caption=None,
             from_data=None, **kwargs
     ):
         super().__init__(connection, Message.Type.file_message,
@@ -1175,6 +1176,7 @@ class FileMessage(Message):
             self._file_path = file_path
             self._thumbnail_content = thumbnail_content
             self._thumbnail_path = thumbnail_path
+            self._caption = caption
         else:
             self._file_content = from_data['file_content']
             self._mime_type = from_data['mime_type']
@@ -1182,6 +1184,7 @@ class FileMessage(Message):
             self._file_path = None
             self._thumbnail_content = from_data['thumbnail_content']
             self._thumbnail_path = None
+            self._caption = from_data['caption']
 
     @property
     def file_content(self):
@@ -1267,6 +1270,10 @@ class FileMessage(Message):
             'i': 0,
         }
 
+        # Add caption (if any)
+        if self._caption:
+            content["d"] = self._caption
+
         # Encrypt and upload thumbnail (if any)
         thumbnail_content = self.thumbnail_content
         if thumbnail_content is not None:
@@ -1313,6 +1320,7 @@ class FileMessage(Message):
         except binascii.Error as exc:
             raise MessageError('Could not convert hex-encoded secret key') from exc
         thumbnail_id = content.get('t')
+        caption = content.get('d')
 
         # Download and decrypt thumbnail (if any)
         if thumbnail_id is not None:
@@ -1339,4 +1347,5 @@ class FileMessage(Message):
             'mime_type': mime_type,
             'file_name': file_name,
             'thumbnail_content': thumbnail_content,
+            'caption': caption,
         }))